pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptforge-openenv"
version = "1.0.0"
description = "PromptForge — OpenEnv-compliant RL environment that trains agents to detect and eliminate Prompt Debt from production LLM system prompts."
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "openai>=1.20.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "gradio>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]
perplexity = [
    "transformers>=4.40.0",
    "torch>=2.2.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["server"]
py-modules = ["models", "client"]
